instruments/requirements.txt
numpy
pyserial
quantities
flufl.enum

